backend/anomaly-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.8.0
python-dotenv==1.0.1
requests>=2.32.0,<3
langchain>=0.2.14,<0.4
langchain-groq>=0.1.9,<0.3
